Update openemr_track_anything_labdata_demo.sql
What is you idea of the last part: procedure_results? Should this be in sync with International values or will this be corrected fist to have the same Unit name as does exist in ..... List ... options? I can imagine there are more values and in Holland the CC is getting uncommon and being replaced by ml.

Are you able to check ip my Translation suggestions are in the right place and if more is needed to do in this File? The last Paragraph need some more attention and I will leave it till next time.

What is pg?